fix:폼제출 controller 및 prechat controller 수정
